+++ src/test/java/org/folio/entitlement/integration/kafka/CapabilitiesModuleEventPublisherTest.java
@@ -21,6 +21,7 @@ import static org.folio.entitlement.support.TestConstants.FLOW_ID;
 import static org.folio.entitlement.support.TestConstants.FLOW_STAGE_ID;
 import static org.folio.entitlement.support.TestConstants.TENANT_ID;
 import static org.folio.entitlement.support.TestConstants.TENANT_NAME;
+import static org.folio.entitlement.support.TestConstants.capabilitiesTenantCollectionTopic;
 import static org.folio.entitlement.support.TestConstants.capabilitiesTenantTopic;
 import static org.folio.entitlement.support.TestUtils.readCapabilityEvent;
 import static org.folio.entitlement.support.TestUtils.readModuleDescriptor;
@@ -30,6 +31,7 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.when;
 
 import java.util.HashMap;
 import java.util.List;
@@ -37,6 +39,7 @@ import java.util.Map;
 import java.util.stream.Stream;
 import org.folio.common.domain.model.ModuleDescriptor;
 import org.folio.entitlement.domain.model.EntitlementRequest;
+import org.folio.entitlement.integration.kafka.configuration.TenantEntitlementKafkaProperties;
 import org.folio.entitlement.integration.kafka.model.CapabilityEventPayload;
 import org.folio.entitlement.integration.kafka.model.ModuleType;
 import org.folio.entitlement.integration.kafka.model.ResourceEvent;
@@ -61,6 +64,7 @@ class CapabilitiesModuleEventPublisherTest {
 
   @InjectMocks private CapabilitiesModuleEventPublisher moduleEventPublisher;
   @Mock private KafkaEventPublisher kafkaEventPublisher;
+  @Mock private TenantEntitlementKafkaProperties tenantEntitlementKafkaProperties;
   @Captor private ArgumentCaptor<ResourceEvent<CapabilityEventPayload>> eventCaptor;
   @Captor private ArgumentCaptor<String> messageKeyCaptor;
 
@@ -81,6 +85,27 @@ class CapabilitiesModuleEventPublisherTest {
 
     var topicName = capabilitiesTenantTopic();
     doNothing().when(kafkaEventPublisher).send(eq(topicName), messageKeyCaptor.capture(), eventCaptor.capture());
+    when(tenantEntitlementKafkaProperties.isProducerTenantCollection()).thenReturn(false);
+
+    moduleEventPublisher.execute(stageContext);
+
+    assertThat(eventCaptor.getAllValues()).containsExactlyElementsOf(expectedEvents);
+    assertThat(messageKeyCaptor.getAllValues()).containsOnly(TENANT_ID.toString());
+  }
+
+  @DisplayName("execute_positive_parameterized_useTenantCollectionTopic")
+  @MethodSource("executeDatasetProvider")
+  @ParameterizedTest(name = "[{index}] {0}")
+  void execute_positive_parameterized_useTenantCollectionTopic(@SuppressWarnings("unused") String testName,
+    ModuleType moduleType, ModuleDescriptor descriptor, ModuleDescriptor installedModuleDescriptor,
+    List<ResourceEvent<CapabilityEventPayload>> expectedEvents) {
+    var contextData = Map.of(PARAM_TENANT_NAME, TENANT_NAME);
+    var flowParameters = flowParameters(moduleType, descriptor, installedModuleDescriptor);
+    var stageContext = moduleStageContext(FLOW_STAGE_ID, flowParameters, contextData);
+
+    var topicName = capabilitiesTenantCollectionTopic();
+    doNothing().when(kafkaEventPublisher).send(eq(topicName), messageKeyCaptor.capture(), eventCaptor.capture());
+    when(tenantEntitlementKafkaProperties.isProducerTenantCollection()).thenReturn(true);
 
     moduleEventPublisher.execute(stageContext);
 
@@ -94,8 +119,9 @@ class CapabilitiesModuleEventPublisherTest {
     var descriptor = readModuleDescriptor("json/events/capabilities/module-desc-with-nothing-to-publish.json");
     var contextData = Map.of(PARAM_TENANT_NAME, TENANT_NAME);
     var flowParameters = moduleFlowParameters(request, descriptor);
-
     var stageContext = moduleStageContext(FLOW_STAGE_ID, flowParameters, contextData);
+
+    when(tenantEntitlementKafkaProperties.isProducerTenantCollection()).thenReturn(false);
 
     moduleEventPublisher.execute(stageContext);
 
@@ -116,6 +142,7 @@ class CapabilitiesModuleEventPublisherTest {
 
     var topicName = capabilitiesTenantTopic();
     doNothing().when(kafkaEventPublisher).send(eq(topicName), messageKeyCaptor.capture(), eventCaptor.capture());
+    when(tenantEntitlementKafkaProperties.isProducerTenantCollection()).thenReturn(false);
 
     moduleEventPublisher.execute(stageContext);
 
@@ -149,6 +176,13 @@ class CapabilitiesModuleEventPublisherTest {
     var result = moduleEventPublisher.getStageName(stageContext);
 
     assertThat(result).isEqualTo("mod-foo-1.0.0-capabilitiesModuleEventPublisher");
+  }
+
+  @Test
+  void getTopicNameByTenantCollection_positive() {
+    var actual = moduleEventPublisher.getTopicNameByTenantCollection();
+
+    assertThat(actual).isEqualTo("folio.ALL.mgr-tenant-entitlements.capability");
   }
 
   private static EntitlementRequest entitlementRequest() {