+++ NEWS.md
@@ -9,6 +9,7 @@
 * Use SECURE_STORE_ENV, not ENV, for secure store key (MGRENTITLE-139)
 * Add support for custom Keycloak base URL for JWKS endpoint, new ENV variable `KC_JWKS_BASE_URL` (MODSIDECAR-148)
 * Implement automatic token refresh for long-running operations to prevent token expiration issues (MGRENTITLE-141)
+* Implement Kafka Tenant Collection Topics (MGRENTITLE-41)
 
 ---